Add default labels and just run with them if we fail to find them via dbus look-up (avoid restart loops)
Commented out the getThumbnail logging for now as they may not exist if the tracking event is old
(and panic as a result)

Added support for null species in the AT ESL interface - like 'any' but assumes 0% threshold
(we may update this to a saner default)

Also added some clip age logging - not yet enabled but starting to focus on the age of the event
to set the AT ESL eventdate (currently assumes it's 'now'). This became evident with the panic's /
restarts from label and thumbnail issues but worth setting the value sensibly for precision.